pyproject.toml
[project]
name = "pixie-qa"
version = "0.4.0"
description = "Automated quality assurance for AI applications"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = ["llm", "evals", "testing", "ai", "observability", "opentelemetry"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-api>=1.27.0",
    "openinference-instrumentation>=0.1.44",
    "autoevals>=0.1.0",
    "pydantic>=2.0",
    "jsonpickle>=4.0.0",
    "python-dotenv>=1.2.2",
    "openai>=2.29.0",
    "starlette>=1.0.0",
    "uvicorn>=0.42.0",
    "watchfiles>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/yiouli/pixie-qa"
Repository = "https://github.com/yiouli/pixie-qa"
Documentation = "https://yiouli.github.io/pixie-qa/"
"Bug Tracker" = "https://github.com/yiouli/pixie-qa/issues"

[project.optional-dependencies]
openai = ["openinference-instrumentation-openai"]
anthropic = ["openinference-instrumentation-anthropic"]
langchain = ["openinference-instrumentation-langchain"]
google = ["openinference-instrumentation-google-genai"]
dspy = ["openinference-instrumentation-dspy"]
all = [
    "openinference-instrumentation-openai",
    "openinference-instrumentation-anthropic",
    "openinference-instrumentation-langchain",
    "openinference-instrumentation-google-genai",
    "openinference-instrumentation-dspy",
]

[project.scripts]
pixie = "pixie.cli.main:main"
pixie-test = "pixie.cli.test_command:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "mypy>=1.10",
    "ruff>=0.4",
    "pytest-asyncio>=1.3.0",
    "pdoc>=14.0",
    "playwright>=1.58.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["pixie"]

[tool.hatch.build.targets.sdist.force-include]
"pixie/assets/webui.html" = "pixie/assets/webui.html"

[tool.hatch.build.targets.wheel]
packages = ["pixie"]

[tool.hatch.build.targets.wheel.force-include]
"pixie/assets/webui.html" = "pixie/assets/webui.html"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "autoevals.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jsonpickle.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "watchfiles.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pixie.evals.scorers"
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
# autoevals has no py.typed; subclassing its types yields spurious [misc] errors
module = "tests.pixie.evals.test_scorers"
disable_error_code = ["import-untyped", "misc"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["UP017"]

[tool.ruff.lint.per-file-ignores]
"scripts/generate_skill_docs.py" = ["E501"]
"pixie/evals/scorers.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests/pixie"]
